requirements.txt
setuptools>=63.4.1
pyyaml>=6.0
opencv-python>=4.6.0.66
numpy>=1.23.4
librosa>=0.9.2
pyyaml>=6.0
scikit-image>=0.19.2
matplotlib==3.5.2
scikit-learn>=1.1.2
mosaic>=1.0
pandas>=1.4.4
torch>=1.13.0
torchvision>=0.13.1
lightning>=2.0.3
pytorch-lightning>=2.0.3
wav2clip>=0.1.0
transformers>=4.30.2
tensorboard>=2.13.0
pillow>=9.5.0
